add some style in css, remove Card.js
