+++ client/src/pages/Resources/Timetable/Registered/Cards.json
@@ -3,31 +3,36 @@
         "title": "Early routine cards",
         "text": "<p>Suitable for age: 2 – 4</p><p>Quantity: 40</p>",
         "src": "~/assets/images/timetable-2-4age.jpg",
-        "alt": "Timetable for age from 2 years to 4 years"
+        "alt": "Timetable for age from 2 years to 4 years",
+        "pdflink": "https://drive.google.com/drive/folders/1ndM1WyLegxUu8kpSJ-Nth8rk4onKiR9K"
     },
      {
         "title": "Homework cards",
         "text": "<p>Suitable for age: 3 – 5</p><p>Quantity: 54</p>",
         "src": "~/assets/images/timetable-3-5age.jpg",
-        "alt": "Timetable for age from 3 years to 5 years"
+        "alt": "Timetable for age from 3 years to 5 years",
+        "pdflink": "https://drive.google.com/drive/folders/1ndM1WyLegxUu8kpSJ-Nth8rk4onKiR9K"
     },
      {
         "title": "Organisational cards",
         "text": "<p>Suitable for age: 7 – 11</p><p>Quantity: 54</p>",
         "src": "~/assets/images/timetable7-11age.jpg",
-        "alt": "Timetable for age from 7 years to 11 years"
+        "alt": "Timetable for age from 7 years to 11 years",
+        "pdflink": "https://drive.google.com/drive/folders/1ndM1WyLegxUu8kpSJ-Nth8rk4onKiR9K"
     },
      {
         "title": "Organisational cards",
         "text": "<p>Suitable for age: 9 – 14</p><p>Quantity: 54</p>",
         "src": "~/assets/images/timetable-9-14.jpg",
-        "alt": "Timetable for age from 9 years to 14 years"
+        "alt": "Timetable for age from 9 years to 14 years",
+        "pdflink": "https://drive.google.com/drive/folders/1ndM1WyLegxUu8kpSJ-Nth8rk4onKiR9K"
     },
      {
         "title": "Mix & Match routine cards",
         "text": "<p>Suitable for all ages</p>",
         "src": "~/assets/images/timetable-mix-match-routine.jpg",
-        "alt": "Timetable for mix match routine"
+        "alt": "Timetable for mix match routine",
+        "pdflink": "https://drive.google.com/drive/folders/1ndM1WyLegxUu8kpSJ-Nth8rk4onKiR9K"
     }
    
 ]
